Refactor Weaviate integration: move health check and add new test endpoint
