+++ fastapi/api/tour_guides/tour_guide_routes.py
@@ -120,23 +120,3 @@ async def get_tour_guides(api_keys=Depends(get_token_then_APIS_cached)):
         logger.error(f"Error retrieving tour guide information: {e}")
         raise HTTPException(status_code=500, detail=str(e))
 
-
-@router.get("/test-weaviate")
-async def test_weaviate_connection(api_keys=Depends(get_token_then_APIS_cached)):
-    """Test the connection to Weaviate using user-specific credentials (with caching)."""
-    weaviate_url = api_keys["weaviate_url"]
-    weaviate_api_key = api_keys["weaviate_api_key"]
-    openai_api_key = api_keys["openai_api_key"]
-    user_id = api_keys["user_id"]
-    try:
-        client = get_weaviate_client(weaviate_url, weaviate_api_key, openai_api_key, user_id)
-        # Try to list collections
-        collections = client.collections.list_all()
-        return {
-            "status": "success",
-            "message": "Successfully connected to Weaviate",
-            "collections": collections
-        }
-    except Exception as e:
-        logger.error(f"Error connecting to Weaviate: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
